Added Functions
Streamlined the app, removed some functions from the edit form.
No need for next or previous card because when playing you wont know which one you want to edit 90% of the time so the next selector is useless. The  number id's and unique names solve that problem.

Added
Overlay for Loading and Edit card
Animation function for Initiative Roll

Functions
Now I can properly retrieve data for the edit form and toggle the select icon.

Bug
Removed bug where Initiative was set as an empty string instead of 0.